:adhesive_bandage: Fixed bugs after merge.

+++ src/main/java/com/schnabel/schnabel/users/repository/ISystemAdminRepository.java
@@ -3,5 +3,6 @@ package com.schnabel.schnabel.users.repository;
 import com.schnabel.schnabel.users.model.SystemAdmin;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ISystemAdminRepository extends CrudRepository<SystemAdmin, Integer> {
+public interface ISystemAdminRepository extends CrudRepository<SystemAdmin, Long>
+{
 }
