+++ src/main/java/com/schnabel/schnabel/users/service/ISystemAdminService.java
@@ -3,5 +3,6 @@ package com.schnabel.schnabel.users.service;
 import com.schnabel.schnabel.misc.interfaces.ICrudService;
 import com.schnabel.schnabel.users.model.SystemAdmin;
 
-public interface ISystemAdminService extends ICrudService<SystemAdmin, Integer> {
+public interface ISystemAdminService extends ICrudService<SystemAdmin, Long>
+{
 }
