+++ src/main/java/com/schnabel/schnabel/users/service/SystemAdminService.java
@@ -4,8 +4,9 @@ import com.schnabel.schnabel.misc.implementations.CrudService;
 import com.schnabel.schnabel.users.model.SystemAdmin;
 import org.springframework.data.repository.CrudRepository;
 
-public class SystemAdminService extends CrudService<SystemAdmin, Integer> implements ISystemAdminService {
-    public SystemAdminService(CrudRepository<SystemAdmin, Integer> repository) {
+public class SystemAdminService extends CrudService<SystemAdmin, Long> implements ISystemAdminService
+{
+    public SystemAdminService(CrudRepository<SystemAdmin, Long> repository) {
         super(repository);
     }
 }
